+++ src/napari_n2v/utils/__init__.py
@@ -3,7 +3,6 @@ from .n2v_utils import (
     create_model,
     filter_dimensions,
     are_axes_valid,
-    build_modelzoo,
     reshape_data,
     get_size_from_shape,
     get_images_count,
@@ -23,9 +22,10 @@ from .io_utils import (
     load_configuration,
     save_configuration,
     load_model,
-    save_model
+    save_model,
+    build_modelzoo
 )
-from .n2v_utils import cwd, State, UpdateType, ModelSaveMode
+from .n2v_utils import cwd, State, UpdateType, ModelSaveMode, Algorithm
 from .n2v_utils import PREDICT, DENOISING, REF_AXES, SAMPLE, NAPARI_AXES
 from .prediction_worker import prediction_after_training_worker, prediction_worker
 from .training_worker import train_worker
