+++ src/napari_n2v/utils/n2v_utils.py
@@ -14,8 +14,7 @@ from napari.utils import notifications as ntf
 
 from n2v.models import N2V, N2VConfig
 
-from napari_n2v.resources import DOC_BIOIMAGE
-from .expert_settings import get_default_settings
+from .expert_settings import get_default_settings, PixelManipulator
 
 REF_AXES = 'TSZYXC'
 NAPARI_AXES = 'TSZYXC'
@@ -23,6 +22,12 @@ NAPARI_AXES = 'TSZYXC'
 PREDICT = '_denoised'
 DENOISING = 'Denoised'
 SAMPLE = 'Example data'
+
+
+class Algorithm(Enum):
+    N2V = 0
+    StructN2V = 1
+    N2V2 = 2
 
 
 class State(Enum):
@@ -49,6 +54,22 @@ class ModelSaveMode(Enum):
     @classmethod
     def list(cls):
         return list(map(lambda c: c.value, cls))
+
+
+def which_algorithm(model: N2V):
+    """
+    Checks which algorithm the model is configured for (N2V, N2V2, structN2V)
+    """
+    # extract configuration
+    config = model.config
+
+    if config.structN2Vmask is not None:
+        return Algorithm.StructN2V
+    elif config.n2v_manipulator == PixelManipulator.MEDIAN.value and \
+            config.unet_residuals and config.blurpool and config.skip_skipone:
+        return Algorithm.N2V2
+    else:
+        return Algorithm.N2V
 
 
 def create_config(X_patches,
@@ -160,43 +181,6 @@ def are_axes_valid(axes: str):
 
     # prior: X and Y contiguous (#FancyComments)
     return ('XY' in _axes) or ('YX' in _axes)
-
-
-def build_modelzoo(path: Union[str, Path], weights: str, inputs, outputs, tf_version: str, axes='byxc'):
-    import os
-    from bioimageio.core.build_spec import build_model
-
-    assert path.endswith('.bioimage.io.zip'), 'Path must end with .bioimage.io.zip'
-
-    tags_dim = '3d' if len(axes) == 5 else '2d'
-    doc = DOC_BIOIMAGE
-
-    head, _ = os.path.split(weights)
-    head = os.path.join(os.path.normcase(head), "config.json")
-    build_model(weight_uri=weights,
-                test_inputs=[inputs],
-                test_outputs=[outputs],
-                input_axes=[axes],
-                output_axes=[axes],
-                output_path=path,
-                name='Noise2Void',
-                description='Self-supervised denoising.',
-                authors=[{'name': "Tim-Oliver Buchholz"}, {'name': "Alexander Krull"}, {'name': "Florian Jug"}],
-                license="BSD-3-Clause",
-                documentation=os.path.abspath(doc),
-                tags=[tags_dim, 'tensorflow', 'unet', 'denoising'],
-                cite=[{'text': 'Noise2Void - Learning Denoising from Single Noisy Images',
-                       'doi': "10.48550/arXiv.1811.10980"}],
-                preprocessing=[[{
-                    "name": "zero_mean_unit_variance",
-                    "kwargs": {
-                        "axes": "yx",
-                        "mode": "per_dataset"
-                    }
-                }]],
-                tensorflow_version=tf_version,
-                attachments={"files": head}
-                )
 
 
 def list_diff(l1, l2):
